+++ libraries/transformer/src/androidTest/java/androidx/media3/transformer/MediaProjectionAssetLoaderTest.java
@@ -42,6 +42,7 @@ import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.ConditionVariable;
 import androidx.media3.common.util.Util;
 import androidx.media3.effect.Presentation;
+import androidx.media3.muxer.Muxer;
 import androidx.media3.muxer.MuxerException;
 import androidx.test.core.app.ActivityScenario;
 import androidx.test.ext.junit.runners.AndroidJUnit4;